Fix compilation on Windows

+++ v8binding/callback_internal.h
@@ -263,7 +263,7 @@ struct V8FunctionInvoker<void(ArgTypes...)> {
     auto context = func->CreationContext();
     v8::Context::Scope context_scope(context);
     std::vector<v8::Local<v8::Value>> args = { ConvertToV8(isolate, raw)... };
-    func->Call(func, args.size(), &args.front());
+    func->Call(func, static_cast<int>(args.size()), &args.front());
   }
 };
 
@@ -282,7 +282,8 @@ struct V8FunctionInvoker<ReturnType(ArgTypes...)> {
     v8::Context::Scope context_scope(context);
     std::vector<v8::Local<v8::Value>> args = { ConvertToV8(isolate, raw)... };
     v8::Local<v8::Value> val;
-    if (func->Call(context, func, args.size(), &args.front()).ToLocal(&val))
+    if (func->Call(context, func, static_cast<int>(args.size()),
+                   &args.front()).ToLocal(&val))
       FromV8(context, val, &ret);
     return ret;
   }
